1sr run

+++ solar_input.py
@@ -7,9 +7,7 @@ from solar_objects import Star, Planet
 def read_space_objects_data_from_file(input_filename):
     """Cчитывает данные о космических объектах из файла, создаёт сами объекты
     и вызывает создание их графических образов
-
     Параметры:
-
     **input_filename** — имя входного файла
     """
 
@@ -38,16 +36,12 @@ def parse_star_parameters(line, star):
     """Считывает данные о звезде из строки.
     Входная строка должна иметь слеюущий формат:
     Star <радиус в пикселах> <цвет> <масса> <x> <y> <Vx> <Vy>
-
     Здесь (x, y) — координаты зведы, (Vx, Vy) — скорость.
     Пример строки:
     Star 10 red 1000 1 2 3 4
-
     Параметры:
-
     **line** — строка с описание звезды.
     **star** — объект звезды.
-
     Planet 2 orange 3.302E23 57.909E9 0 0 47.87E3
     """
 
@@ -56,32 +50,30 @@ def parse_star_parameters(line, star):
 
     splitted = line.split()
 
-    star.m = int(splitted[3])
+    star.m = float(splitted[3])
 
-    star.x = int(splitted[4])
+    star.x = float(splitted[4])
 
-    star.y = int(splitted[5])
+    star.y = float(splitted[5])
 
-    star.Vx = int(splitted[6])
+    star.Vx = float(splitted[6])
 
-    star.Vy = int(splitted[7])
+    star.Vy = float(splitted[7])
 
-    star.r = int(splitted[1])
+    star.r = float(splitted[1])
 
     star.color = splitted[2]
+
 
 def parse_planet_parameters(line, planet):
     """Считывает данные о планете из строки.
     Предполагается такая строка:
     Входная строка должна иметь слеюущий формат:
     Planet <радиус в пикселах> <цвет> <масса> <x> <y> <Vx> <Vy>
-
     Здесь (x, y) — координаты планеты, (Vx, Vy) — скорость.
     Пример строки:
     Planet 10 red 1000 1 2 3 4
-
     Параметры:
-
     **line** — строка с описание планеты.
     **planet** — объект планеты.
     """
@@ -90,17 +82,17 @@ def parse_planet_parameters(line, planet):
 
     splitted = line.split()
 
-    planet.m = int(splitted[3])
+    planet.m = float(splitted[3])
 
-    planet.x = int(splitted[4])
+    planet.x = float(splitted[4])
 
-    planet.y = int(splitted[5])
+    planet.y = float(splitted[5])
 
-    planet.Vx = int(splitted[6])
+    planet.Vx = float(splitted[6])
 
-    planet.Vy = int(splitted[7])
+    planet.Vy = float(splitted[7])
 
-    planet.r = int(splitted[1])
+    planet.r = float(splitted[1])
 
     planet.color = splitted[2]
 
@@ -110,9 +102,7 @@ def write_space_objects_data_to_file(output_filename, space_objects):
     Строки должны иметь следующий формат:
     Star <радиус в пикселах> <цвет> <масса> <x> <y> <Vx> <Vy>
     Planet <радиус в пикселах> <цвет> <масса> <x> <y> <Vx> <Vy>
-
     Параметры:
-
     **output_filename** — имя входного файла
     **space_objects** — список объектов планет и звёзд
     """
